Fix: basic logic (not using graphql)
