+++ spring-for-graphql/src/main/java/com/example/demo/controller/MemberGraphQLController.java
@@ -1,11 +1,7 @@
-package com.example.demo.member.controller;
+package com.example.demo.controller;
 
-import com.example.demo.member.Member;
-import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/graphql")
