+++ spring-for-graphql/src/main/java/com/example/demo/domain/MemberRole.java
@@ -1,4 +1,4 @@
-package com.example.demo.member;
+package com.example.demo.domain;
 
 
 public enum MemberRole {
