+++ spring-for-graphql/src/main/java/com/example/demo/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.example.demo.member;
+package com.example.demo.repository;
 
+import com.example.demo.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
