+++ spring-for-graphql/src/main/resources/application.yml
@@ -2,9 +2,12 @@ spring:
   graphql:
     graphiql:
       enabled: true
+      # graphiql을 true를 설정해 주면 localhost:8080/graphiql (Default path)을 통해 graphql 쿼리를 테스트가 가능합니다.
+      # 이 방법 말고도 IntelliJ에 GraphQL 플러그인을 설치해서 IntelliJ에서 직접 테스트도 가능하며 Postman으로도 가능합니다.
     schema:
       printer:
         enabled: true
+        # 이 설정 시 jpa 의 show-sql 같이 graphql 쿼리를 출력해 줍니다.
   h2:
     console:
       path: "/h2-console"
@@ -15,4 +18,4 @@ spring:
     driver-class-name: org.h2.Driver
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: create-drop
